chart nav and change price to Long

+++ modules/domain/src/main/java/com/eastbarnetschool/ordermatchingengine/domain/Order.java
@@ -1,14 +1,14 @@
 package com.eastbarnetschool.ordermatchingengine.domain;
 
 public class Order {
-    private Double price;
+    private Long price;
     private Integer initialQuantity;
     private Integer remainingQuantity;
     private String ticker;
     private Side side;
     private OrderType orderType;
 
-    public Order(Double price, Integer quantity, String ticker, Side side, OrderType orderType) {
+    public Order(Long price, Integer quantity, String ticker, Side side, OrderType orderType) {
         this.price = price;
         this.initialQuantity = quantity;
         this.remainingQuantity = quantity;
@@ -17,7 +17,7 @@ public class Order {
         this.orderType = orderType;
     }
 
-    public Double getPrice() {
+    public Long getPrice() {
         return price;
     }
