+++ modules/domain/src/main/java/com/eastbarnetschool/ordermatchingengine/domain/Trade.java
@@ -4,9 +4,9 @@ import java.time.Instant;
 
 public class Trade {
     private Instant timestamp;
-    private Double price;
+    private Long price;
     private Integer quantity;
-    public Trade(Instant timestamp, Double price, Integer quantity) {
+    public Trade(Instant timestamp, Long price, Integer quantity) {
         this.timestamp = timestamp;
         this.price = price;
         this.quantity = quantity;
